Split MenuItem components into separate files
