+++ lib/components/MenuItem/MenuItemDivider.tsx
@@ -1,0 +1,8 @@
+import React from 'react';
+import { Box, Divider } from '../';
+
+export const MenuItemDivider = () => (
+  <Box paddingX="xxsmall" paddingY="xxsmall">
+    <Divider />
+  </Box>
+);
