+++ lib/components/MenuItem/useMenuItem.treat.ts
@@ -1,0 +1,10 @@
+import { style } from 'sku/treat';
+
+export const menuItem = style({
+  whiteSpace: 'nowrap',
+  selectors: {
+    [`&::-moz-focus-inner`]: {
+      border: 0,
+    },
+  },
+});
